fix client server boundry for uploads
